# coordinator: Dequeue batches of messages