Dequeues batches of messages at a time when possible to amortize the cost
of calling into select. Not all components are able to provide batches of
messages, specifically the controller-ready response is message-at-a-time.

Workload: (The exact view definition is pretty irrelevant, just a leftover from a previous experiment.)
```
duration = 120s

[setup]
query=create materialized view if not exists test2 as select generate_series as x from generate_series(1, 100)
query=create view if not exists tempview1 as select * from test2 where x % 2 = 1
query=create default index on tempview1

[teardown]
query=drop materialized view test2 cascade

[loadtest]
query=select * from tempview1
concurrency=128
start=5s
```

Before:
```
2024/08/15 13:33:23 loadtest: 167229 transactions (1453.389 TPS), latency 87.544447ms±329.584µs; 8361450 rows (72669.429 RPS), 167229 queries (1453.389 QPS); 0 aborts (0.000%), latency 0s±0s
Transactions:
  8.388608ms -  16.777216ms [     1]: ▏
 16.777216ms -  33.554432ms [     0]:
 33.554432ms -  67.108864ms [ 92177]: ██████████████████████████████████████████████████
 67.108864ms - 134.217728ms [ 49124]: ██████████████████████████▋
134.217728ms - 268.435456ms [ 23421]: ████████████▋
268.435456ms - 536.870912ms [  2375]: █▎
536.870912ms - 1.073741824s [   131]: ▏
2024/08/15 13:33:23 Performing teardown
```

After:
```
2024/08/16 10:20:30 loadtest: 230644 transactions (2004.556 TPS), latency 63.369413ms±183.222µs; 11532200 rows (100227.795 RPS), 230644 queries (2004.556 QPS); 0 aborts (0.000%), latency 0s±0s
Transactions:
 16.777216ms -  33.554432ms [   669]: ▏
 33.554432ms -  67.108864ms [161896]: ██████████████████████████████████████████████████
 67.108864ms - 134.217728ms [ 58666]: ██████████████████
134.217728ms - 268.435456ms [  8837]: ██▋
268.435456ms - 536.870912ms [   576]: ▏
2024/08/16 10:20:30 Performing teardown
```

In support of #28720.

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly considered. ([trigger-ci for additional test/nightly runs](https://trigger-ci.dev.materialize.com/))
- [ ] This PR has an associated up-to-date [design doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md), is a design doc ([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)), or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T` mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md) (possibly in a backwards-incompatible way), then it is tagged with a `T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests, there is a companion cloud PR to account for those changes that is tagged with the release-blocker label ([example](https://github.com/MaterializeInc/cloud/pull/5021)).
  <!-- Ask in #team-cloud on Slack if you need help preparing the cloud PR. -->
- [ ] If this PR includes major [user-facing behavior changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note), I have pinged the relevant PM to schedule a changelog post.