@antiguru 2024-08-16
Moving the `boxed_local` in `handle_message` into each await doesn't have a large effect:
```
2024/08/16 11:31:55 loadtest: 238885 transactions (2076.805 TPS), latency 61.158558ms±175.072µs; 11944250 rows (103840.268 RPS), 238885 queries (2076.805 QPS); 0 aborts (0.000%), latency 0s±0s
Transactions:
  8.388608ms -  16.777216ms [     1]: ▏
 16.777216ms -  33.554432ms [  1238]: ▎
 33.554432ms -  67.108864ms [168607]: ██████████████████████████████████████████████████
 67.108864ms - 134.217728ms [ 59563]: █████████████████▋
134.217728ms - 268.435456ms [  9263]: ██▋
268.435456ms - 536.870912ms [   213]: ▏
2024/08/16 11:31:55 Performing teardown
```