@benesch 2024-08-16
@ParkMyCar's concern around ordering got me thinking. If `select!` is what's expensive, we could fairly easily rewrite the command loop too not use `select!`. There's no reason that we need to conjure up messages from fifteen different streams. We could instead have a single command channel, and various newtypes over `cmd_tx` that are handed out to the components that need to send commands and enforce that each component only sends the type of commands that are appropriate for that component. 

Then the coordinator loop would just be a single `cmd_rx.recv()`. 

There's some trickiness around priority (see the `biased` in today's `select!` macro), but that could easily be resolved by draining all commands in the receiver, sorting them by priority, and then processing them in priority order.