@antiguru 2024-08-20
Nightly tests in https://buildkite.com/materialize/nightly/builds/9194