@antiguru 2024-08-22
Updated to call `advance_timelines` directly.